+++ src/Symfony/Bundle/FrameworkBundle/Controller/AbstractController.php
@@ -19,7 +19,6 @@ use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\Form\FormInterface;
-use Symfony\Component\Form\FormView;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
 use Symfony\Component\HttpFoundation\Exception\SessionNotFoundException;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -28,6 +27,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\ResponseHeaderBag;
+use Symfony\Component\HttpFoundation\Session\FlashBagAwareSessionInterface;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
@@ -173,10 +173,16 @@ abstract class AbstractController implements ServiceSubscriberInterface
     protected function addFlash(string $type, mixed $message): void
     {
         try {
-            $this->container->get('request_stack')->getSession()->getFlashBag()->add($type, $message);
+            $session = $this->container->get('request_stack')->getSession();
         } catch (SessionNotFoundException $e) {
             throw new \LogicException('You cannot use the addFlash method if sessions are disabled. Enable them in "config/packages/framework.yaml".', 0, $e);
         }
+
+        if (!$session instanceof FlashBagAwareSessionInterface) {
+            trigger_deprecation('symfony/framework-bundle', '6.2', 'Calling "addFlash()" method when the session does not implement %s is deprecated.', FlashBagAwareSessionInterface::class);
+        }
+
+        $session->getFlashBag()->add($type, $message);
     }
 
     /**
@@ -212,6 +218,8 @@ abstract class AbstractController implements ServiceSubscriberInterface
 
     /**
      * Returns a rendered view.
+     *
+     * Forms found in parameters are auto-cast to form views.
      */
     protected function renderView(string $view, array $parameters = []): string
     {
@@ -219,11 +227,20 @@ abstract class AbstractController implements ServiceSubscriberInterface
             throw new \LogicException('You cannot use the "renderView" method if the Twig Bundle is not available. Try running "composer require symfony/twig-bundle".');
         }
 
+        foreach ($parameters as $k => $v) {
+            if ($v instanceof FormInterface) {
+                $parameters[$k] = $v->createView();
+            }
+        }
+
         return $this->container->get('twig')->render($view, $parameters);
     }
 
     /**
      * Renders a view.
+     *
+     * If an invalid form is found in the list of parameters, a 422 status code is returned.
+     * Forms found in parameters are auto-cast to form views.
      */
     protected function render(string $view, array $parameters = [], Response $response = null): Response
     {
@@ -231,6 +248,15 @@ abstract class AbstractController implements ServiceSubscriberInterface
 
         if (null === $response) {
             $response = new Response();
+        }
+
+        if (200 === $response->getStatusCode()) {
+            foreach ($parameters as $v) {
+                if ($v instanceof FormInterface && $v->isSubmitted() && !$v->isValid()) {
+                    $response->setStatusCode(422);
+                    break;
+                }
+            }
         }
 
         $response->setContent($content);
@@ -242,28 +268,12 @@ abstract class AbstractController implements ServiceSubscriberInterface
      * Renders a view and sets the appropriate status code when a form is listed in parameters.
      *
      * If an invalid form is found in the list of parameters, a 422 status code is returned.
+     *
+     * @deprecated since Symfony 6.2, use render() instead
      */
     protected function renderForm(string $view, array $parameters = [], Response $response = null): Response
     {
-        if (null === $response) {
-            $response = new Response();
-        }
-
-        foreach ($parameters as $k => $v) {
-            if ($v instanceof FormView) {
-                throw new \LogicException(sprintf('Passing a FormView to "%s::renderForm()" is not supported, pass directly the form instead for parameter "%s".', get_debug_type($this), $k));
-            }
-
-            if (!$v instanceof FormInterface) {
-                continue;
-            }
-
-            $parameters[$k] = $v->createView();
-
-            if (200 === $response->getStatusCode() && $v->isSubmitted() && !$v->isValid()) {
-                $response->setStatusCode(422);
-            }
-        }
+        trigger_deprecation('symfony/framework-bundle', '6.2', 'The "%s::renderForm()" method is deprecated, use "render()" instead.', get_debug_type($this));
 
         return $this->render($view, $parameters, $response);
     }