+++ src/Symfony/Component/HttpKernel/Kernel.php
@@ -17,7 +17,6 @@ use Symfony\Component\Config\Builder\ConfigBuilderGenerator;
 use Symfony\Component\Config\ConfigCache;
 use Symfony\Component\Config\Loader\DelegatingLoader;
 use Symfony\Component\Config\Loader\LoaderResolver;
-use Symfony\Component\Debug\DebugClassLoader as LegacyDebugClassLoader;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\Compiler\PassConfig;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -78,15 +77,15 @@ abstract class Kernel implements KernelInterface, RebootableInterface, Terminabl
      */
     private static array $freshCache = [];
 
-    public const VERSION = '6.1.3-DEV';
-    public const VERSION_ID = 60103;
+    public const VERSION = '6.2.0-DEV';
+    public const VERSION_ID = 60200;
     public const MAJOR_VERSION = 6;
-    public const MINOR_VERSION = 1;
-    public const RELEASE_VERSION = 3;
+    public const MINOR_VERSION = 2;
+    public const RELEASE_VERSION = 0;
     public const EXTRA_VERSION = 'DEV';
 
-    public const END_OF_MAINTENANCE = '01/2023';
-    public const END_OF_LIFE = '01/2023';
+    public const END_OF_MAINTENANCE = '07/2023';
+    public const END_OF_LIFE = '07/2023';
 
     public function __construct(string $environment, bool $debug)
     {
@@ -513,7 +512,7 @@ abstract class Kernel implements KernelInterface, RebootableInterface, Terminabl
 
                 // Remove frames added by DebugClassLoader.
                 for ($i = \count($backtrace) - 2; 0 < $i; --$i) {
-                    if (\in_array($backtrace[$i]['class'] ?? null, [DebugClassLoader::class, LegacyDebugClassLoader::class], true)) {
+                    if (DebugClassLoader::class === ($backtrace[$i]['class'] ?? null)) {
                         $backtrace = [$backtrace[$i + 1]];
                         break;
                     }