+++ src/Symfony/Component/Serializer/DataCollector/SerializerDataCollector.php
@@ -70,68 +70,68 @@ class SerializerDataCollector extends DataCollector implements LateDataCollector
         return $totalTime;
     }
 
-    public function collectSerialize(string $traceId, mixed $data, string $format, array $context, float $time): void
+    public function collectSerialize(string $traceId, mixed $data, string $format, array $context, float $time, array $caller): void
     {
         unset($context[TraceableSerializer::DEBUG_TRACE_ID]);
 
         $this->collected[$traceId] = array_merge(
             $this->collected[$traceId] ?? [],
-            compact('data', 'format', 'context', 'time'),
+            compact('data', 'format', 'context', 'time', 'caller'),
             ['method' => 'serialize'],
         );
     }
 
-    public function collectDeserialize(string $traceId, mixed $data, string $type, string $format, array $context, float $time): void
+    public function collectDeserialize(string $traceId, mixed $data, string $type, string $format, array $context, float $time, array $caller): void
     {
         unset($context[TraceableSerializer::DEBUG_TRACE_ID]);
 
         $this->collected[$traceId] = array_merge(
             $this->collected[$traceId] ?? [],
-            compact('data', 'format', 'type', 'context', 'time'),
+            compact('data', 'format', 'type', 'context', 'time', 'caller'),
             ['method' => 'deserialize'],
         );
     }
 
-    public function collectNormalize(string $traceId, mixed $data, ?string $format, array $context, float $time): void
+    public function collectNormalize(string $traceId, mixed $data, ?string $format, array $context, float $time, array $caller): void
     {
         unset($context[TraceableSerializer::DEBUG_TRACE_ID]);
 
         $this->collected[$traceId] = array_merge(
             $this->collected[$traceId] ?? [],
-            compact('data', 'format', 'context', 'time'),
+            compact('data', 'format', 'context', 'time', 'caller'),
             ['method' => 'normalize'],
         );
     }
 
-    public function collectDenormalize(string $traceId, mixed $data, string $type, ?string $format, array $context, float $time): void
+    public function collectDenormalize(string $traceId, mixed $data, string $type, ?string $format, array $context, float $time, array $caller): void
     {
         unset($context[TraceableSerializer::DEBUG_TRACE_ID]);
 
         $this->collected[$traceId] = array_merge(
             $this->collected[$traceId] ?? [],
-            compact('data', 'format', 'type', 'context', 'time'),
+            compact('data', 'format', 'type', 'context', 'time', 'caller'),
             ['method' => 'denormalize'],
         );
     }
 
-    public function collectEncode(string $traceId, mixed $data, ?string $format, array $context, float $time): void
+    public function collectEncode(string $traceId, mixed $data, ?string $format, array $context, float $time, array $caller): void
     {
         unset($context[TraceableSerializer::DEBUG_TRACE_ID]);
 
         $this->collected[$traceId] = array_merge(
             $this->collected[$traceId] ?? [],
-            compact('data', 'format', 'context', 'time'),
+            compact('data', 'format', 'context', 'time', 'caller'),
             ['method' => 'encode'],
         );
     }
 
-    public function collectDecode(string $traceId, mixed $data, ?string $format, array $context, float $time): void
+    public function collectDecode(string $traceId, mixed $data, ?string $format, array $context, float $time, array $caller): void
     {
         unset($context[TraceableSerializer::DEBUG_TRACE_ID]);
 
         $this->collected[$traceId] = array_merge(
             $this->collected[$traceId] ?? [],
-            compact('data', 'format', 'context', 'time'),
+            compact('data', 'format', 'context', 'time', 'caller'),
             ['method' => 'decode'],
         );
     }
@@ -192,6 +192,7 @@ class SerializerDataCollector extends DataCollector implements LateDataCollector
                 'context' => $this->cloneVar($collected['context']),
                 'normalization' => [],
                 'encoding' => [],
+                'caller' => $collected['caller'] ?? null,
             ];
 
             if (isset($collected['normalization'])) {